requirements.txt
streamlit
langchain
langchain-google-genai
python-dotenv
faiss-cpu
pypdf